CU11. Mostrar lista peticiones alquiler
