--- back/src/main/java/com/tfg/GoAway/modules/favorite/infraestructure/in/http/FavoriteAdvertisementCreatePostResponse.java
@@ -1,15 +1,0 @@
-package com.tfg.GoAway.modules.favorite.infraestructure.in.http;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class FavoriteAdvertisementCreatePostResponse {
-
-    private String userEmail;       
-
-    private String advertisementId; 
-    
-    private String message;         
-}
