+++ back/src/main/java/com/tfg/GoAway/modules/favorite/infrastructure/in/http/FavoriteAdvertisementCreatePostMapper.java
@@ -1,0 +1,16 @@
+package com.tfg.GoAway.modules.favorite.infrastructure.in.http;
+
+import org.springframework.stereotype.Component;
+
+import com.tfg.GoAway.modules.favorite.domain.FavoriteAdvertisement;
+
+@Component
+public class FavoriteAdvertisementCreatePostMapper {
+
+    public FavoriteAdvertisement toDomain(FavoriteAdvertisementCreatePostRequest request, String userEmail) {
+        return FavoriteAdvertisement.builder()
+                .userEmail(userEmail)
+                .advertisementId(request.getAdvertisementId())
+                .build();
+    }
+}
