+++ back/src/main/java/com/tfg/GoAway/modules/favorite/infrastructure/in/http/FavoriteAdvertisementCreatePostRequest.java
@@ -1,0 +1,11 @@
+package com.tfg.GoAway.modules.favorite.infrastructure.in.http;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class FavoriteAdvertisementCreatePostRequest {
+
+    private String advertisementId; // ID del anuncio que se marcará como favorito
+}
