+++ back/src/main/java/com/tfg/GoAway/modules/favorite/infrastructure/in/http/FavoriteAdvertisementCreatePostResponse.java
@@ -1,0 +1,15 @@
+package com.tfg.GoAway.modules.favorite.infrastructure.in.http;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class FavoriteAdvertisementCreatePostResponse {
+
+    private String userEmail;       
+
+    private String advertisementId; 
+    
+    private String message;         
+}
