+++ back/src/main/java/com/tfg/GoAway/modules/transaction/infrastructure/in/http/create/TransactionCreatePostRequest.java
@@ -1,0 +1,17 @@
+package com.tfg.GoAway.modules.transaction.infrastructure.in.http.create;
+
+import lombok.Getter;
+import lombok.Setter;
+import java.time.LocalDate;
+
+@Getter
+@Setter
+public class TransactionCreatePostRequest {
+
+    private LocalDate startDate;
+    private LocalDate endDate;
+    private Double totalPrice;
+    private String paymentMethod;
+    private String advertisementId;
+    private String ownerEmail; 
+}
