+++ back/src/main/java/com/tfg/GoAway/modules/transaction/infrastructure/in/http/find_all_owner/TransactionGetByOwnerController.java
@@ -1,0 +1,27 @@
+package com.tfg.GoAway.modules.transaction.infrastructure.in.http.find_all_owner;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import com.tfg.GoAway.modules.shared.security.SecurityUtils;
+import com.tfg.GoAway.modules.transaction.application.find_all_owner.TransactionGetByOwner;
+import com.tfg.GoAway.modules.transaction.application.find_all_owner.TransactionGetOwnerResponse;
+
+import lombok.RequiredArgsConstructor;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/transactions")
+public class TransactionGetByOwnerController {
+
+    private final TransactionGetByOwner transactionGetByOwner;
+
+    @GetMapping("/owner")
+    public ResponseEntity<List<TransactionGetOwnerResponse>> getTransactionsByOwner() {
+        String ownerEmail = SecurityUtils.getUserEmailFromContext();
+        List<TransactionGetOwnerResponse> transactions = transactionGetByOwner.execute(ownerEmail);
+        return ResponseEntity.ok(transactions);
+    }
+}
