+++ back/src/main/java/com/tfg/GoAway/modules/transaction/infrastructure/in/http/find_all_tenant/TransactionGetByTenantController.java
@@ -1,0 +1,27 @@
+package com.tfg.GoAway.modules.transaction.infrastructure.in.http.find_all_tenant;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import com.tfg.GoAway.modules.shared.security.SecurityUtils;
+import com.tfg.GoAway.modules.transaction.application.find_all_tenant.TransactionGetByTenant;
+import com.tfg.GoAway.modules.transaction.application.find_all_tenant.TransactionGetTenantResponse;
+
+import lombok.RequiredArgsConstructor;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/transactions")
+public class TransactionGetByTenantController {
+
+    private final TransactionGetByTenant transactionGetByTenant;
+
+    @GetMapping("/tenant")
+    public ResponseEntity<List<TransactionGetTenantResponse>> getTransactionsByOwner() {
+        String ownerEmail = SecurityUtils.getUserEmailFromContext();
+        List<TransactionGetTenantResponse> transactions = transactionGetByTenant.execute(ownerEmail);
+        return ResponseEntity.ok(transactions);
+    }
+}
